initial commit; split up train_gpt_mlx.py
