+++ main.py
@@ -1,0 +1,6 @@
+def main():
+    print("Hello from parameter-golf!")
+
+
+if __name__ == "__main__":
+    main()
