+++ pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "parameter-golf"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "datasets>=4.8.4",
+    "huggingface-hub>=1.9.0",
+    "kernels>=0.12.3",
+    "mlx>=0.31.1",
+    "numpy>=2.4.4",
+    "sentencepiece>=0.2.1",
+    "setuptools>=82.0.1",
+    "tiktoken>=0.12.0",
+    "torch>=2.10.0",
+    "tqdm>=4.67.3",
+    "typing-extensions==4.15.0",
+]
