+++ train_gpt_mlx_lib/__init__.py
@@ -1,0 +1,3 @@
+from .runner import main
+
+__all__ = ["main"]
